requirements.txt
fastapi==0.107.0
uvicorn[standard]==0.23.2
pydantic-settings==2.2.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
numpy==1.26.0
python-dotenv==1.0.0
jinja2==3.1.4